delete post

+++ html/delete.php
@@ -14,22 +14,26 @@
     </ul>
 </nav>
 <body>
-    <h1>Delete an item from the inventory</h1>
-    <form action="delete.php" method="<?php echo $_SERVER['PHP_SELF'];?>">
-    <input type = "text" name = "ID" id = "ID"><br>
-    <input type ="submit" name = "Delete Item">
-    </form>
-<?php
-    require "connect.php";
-    $ID = $_POST["ID"];
-    $query = "DELETE FROM Product WHERE barcode='$ID'";
-    $res=mysqli_query($conn, $query);
-    if($res) {
-        echo"<p>'$ID' was successfully deleted</p>";
-    }
-    else{
-        echo"<p>'$ID' failed to be deleted</p>";
-    }
-?>
+    <div class="box">
+        <h1>Delete an item from the inventory</h1>
+        <form action="delete.php" method="<?php echo $_SERVER['PHP_SELF'];?>">
+            <input type = "text" name = "ID" id = "ID"><br>
+            <input type ="submit" name = "Delete Item">
+        </form>
+        <?php
+            if ($_SERVER["REQUEST_METHOD"] == "POST") {
+            require "connect.php";
+            $ID = $_POST["ID"];
+            $query = "DELETE FROM Product WHERE barcode='$ID'";
+            $res=mysqli_query($conn, $query);
+            if($res) {
+                echo"<p>'$ID' was successfully deleted</p>";
+            }
+            else{
+                echo"<p>'$ID' failed to be deleted</p>";
+            }
+            }
+        ?>
+    </div>
 </body>
 </html>
